+++ cmd/write-history.go
@@ -21,6 +21,8 @@ var (
 
 func init() {
 	AddDefaultFlags(writeHistoryCmd)
+	writeHistoryCmd.PersistentFlags().
+		StringP("tagged", "t", "", "Print only write events for resources with the given tag. Example: -t key:value")
 	// Redefine this flag as we want a different default.
 	writeHistoryCmd.PersistentFlags().
 		StringP("since", "s", "24h", "Since flag. Valid time units are 'ns', 'us' (or 'µs'), 'ms', 's', 'm', 'h'.")
@@ -34,13 +36,19 @@ func parseDurationToUTC(input string) (time.Time, error) {
 	return time.Now().UTC().Add(-duration), nil
 }
 
-func runWriteHistory(cmd *cobra.Command, args []string) error {
+func runWriteHistory(cmd *cobra.Command, _ []string) error {
 	since, _ := cmd.Flags().GetString("since")
 	region, _ := cmd.Flags().GetString("region")
 	raw, _ := cmd.Flags().GetBool("raw")
 	ignoredUsersParam, _ := cmd.Flags().GetString("ignore-users")
 	ignoredUsers := strings.Split(ignoredUsersParam, ",")
 	toggleEventID, _ := cmd.Flags().GetBool("toggle-event-ids")
+
+	resourceTag, _ := cmd.Flags().GetString("tagged")
+	var tag *aws.ResourceTag
+	if tagSplit := strings.Split(resourceTag, ":"); len(tagSplit) >= 2 { //nolint:gomnd // it's an obvious split into KV
+		tag = &aws.ResourceTag{Key: tagSplit[0], Value: tagSplit[1]}
+	}
 
 	startTime, err := parseDurationToUTC(since)
 	if err != nil {
@@ -56,18 +64,21 @@ func runWriteHistory(cmd *cobra.Command, args []string) error {
 	if err != nil {
 		return fmt.Errorf("could not get caller identity: %w", err)
 	}
-	fmt.Println(
-		"Checking write event history since",
+
+	informativeString := fmt.Sprintf(
+		"Checking write event history since %s for AWS account %s as %s",
 		startTime,
-		"for AWS account",
 		*callerIdentity.Account,
-		"as",
 		*callerIdentity.Arn,
 	)
+	if tag != nil {
+		informativeString += fmt.Sprintf(" for resources tagged with '%s:%s'", tag.Key, tag.Value)
+	}
+	fmt.Print(informativeString)
 
 	fmt.Println("")
 	fmt.Println("Fetching", awsClient.Region, "events...")
-	err = awsClient.PrintCloudTrailWriteEvents(startTime, raw, ignoredUsers, toggleEventID)
+	err = awsClient.PrintCloudTrailWriteEvents(startTime, raw, ignoredUsers, toggleEventID, tag)
 	if err != nil {
 		return err
 	}
@@ -81,7 +92,7 @@ func runWriteHistory(cmd *cobra.Command, args []string) error {
 			return fmt.Errorf("could not initialize aws client: %w", err)
 		}
 
-		err = awsClient.PrintCloudTrailWriteEvents(startTime, raw, ignoredUsers, toggleEventID)
+		err = awsClient.PrintCloudTrailWriteEvents(startTime, raw, ignoredUsers, toggleEventID, tag)
 		if err != nil {
 			return err
 		}