+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.4] - 2023-10-05
+
+### Added
+
+- Added filtering write events by resource tags
+
 ## [0.0.3] - 2023-10-02
 
 ### Added